api/requirements.txt
openstacksdk
pymongo
flask
pika